Update README.md
I've noticed that the subtitles in the search results do not appear now if you set isNavigableOnly to True, this is a change made on Notions side. For this reason I have set the parameter to default to False. You can still set this to True but you will lose subtitles in your Alfred search results.